+++ cmake/thirdparty/SetupMPI.cmake
@@ -88,6 +88,15 @@ if (ENABLE_FIND_MPI)
             list(APPEND _mpi_link_flags ${MPI_CXX_LINK_FLAGS})
         endif()
     endif()
+    # Fixes for linking with NVCC
+    if (CUDA_LINK_WITH_NVCC)
+        # Convert rpath flag if linking with CUDA
+        string(REPLACE "-Wl,-rpath," "-Xlinker -rpath -Xlinker "
+                       _mpi_link_flags "${_mpi_link_flags}")
+        # -pthread just doesn't work with nvcc                       
+        string(REPLACE "-pthread" " "
+                       _mpi_link_flags "${_mpi_link_flags}")
+    endif()
 
     # Libraries
     set(_mpi_libraries ${MPI_C_LIBRARIES} ${MPI_CXX_LIBRARIES})